% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvianFlu.R
\docType{package}
\name{AvianFlu}
\alias{AvianFlu}
\alias{AvianFlu-package}
\title{Detections of Avian Flu}
\description{
This package contains descriptions of detections of Avian Flu
as recorded by the Animal and Plant Health Inspection Service
in 2022-2023
}
\seealso{
\link{Mammals}, \link{Wild_Birds}
}
