% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Mammals}
\alias{Mammals}
\title{Mammals Data}
\format{
A data frame with 206 rows and 6 columns:
\describe{
\item{state}{State name}
\item{county}{County name}
\item{date_collected}{Date collected (Year/Month/Day)}
\item{date_detected}{indicates the date when a positive detection was obtained by the developmental RRT PCR (Year/Month/Day)}
\item{hpai_strain}{Name of strain}
\item{species}{Species strain was detected in}
}
}
\source{
\url{https://www.aphis.usda.gov/aphis/ourfocus/animalhealth/animal-disease-information/avian/avian-influenza/hpai-2022/2022-hpai-mammals}
}
\usage{
Mammals
}
\description{
A dataset made by APHIS containing detections of avian influenza in Mammals
from 2022 to 2023.
}
\keyword{datasets}
